+++ src/auth/socket/index.ts
@@ -1,3 +1,4 @@
 export * from './socket-state';
 export * from './socket-token-provider';
 export * from './socket-coordinator';
+export * from './socket-tab-coordinator';